stock_service/requirements.txt
gevent
grpcio
grpcio-tools
pymongo
numpy
pandas
xlrd
